cleaned main.js by putting functions in diff files
